requirements.txt
Django==4.1.4
mysqlclient==2.1.1
selenium==3.141.0
jieba==0.42.1
baidu-aip==4.16.7